ci: run scale census smoke after build
